refactor(context): split the store module along its existing seams
Pure move, no behavior change — the first commit of #712, because every
deliverable after it grows a file that was already 2,224 lines and
grandfathered in the size ratchet.

The seams were already there; this makes them modules:

  store/schema.rs     DDL, the version ladder, pragmas, fingerprint registry
  store/node.rs       the typed vocabulary, write shape, read shape, statements
  store/edge.rs       assertion, supersession, the two point-in-time readers
  store/embedding.rs  the vector codec and the fingerprinted index reads
  store/record.rs     episode and memory rows
  store/domain.rs     the tag table, its junctions, the scope anti-join
  store/tests.rs      the 816-line inline test module, verbatim

`store.rs` keeps `ContextStore` itself and re-exports every moved name,
so `crate::store::X` still resolves for all five consumers and nothing
outside this directory changed. Visibility widened only where the split
demanded it (`pub(crate)` on items two modules now share).

The store's baseline entry is retired rather than lowered: 2,224 lines
becomes 369, which is under the 1,500 limit, and an exemption must not
outlive the problem it covered.

73 stella-context tests pass unchanged.

Part of #712.
